requirements.txt
Django>=4.2
psycopg2-binary>=2.9
redis>=5.0
django-redis>=5.4
gunicorn
Pillow>=10.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3.0
django-filter>=24.3
whitenoise>=6.7
requests>=2.32
dj-database-url>=2.1.0
django-cors-headers>=4.3.0
